[Base/POSIX] Build the thread affinity mask with a 64-bit shift
set_affinity_mask tested `mask & (1 << i)` for i up to 63. `1` is an int, so the
shift is undefined from i == 31 and wraps to the low 32 bits above it: every CPU
index from 32 up was selected from the wrong mask bit, and CPUs 0-31 picked up
spurious partners.

Measured with the two loops side by side:

  mask = CPU 0 only     old selects CPUs 0 and 32   new selects 0
  mask = CPU 33 only    old selects CPUs 31 and 63  new selects 33
  mask = CPUs 32-35     old selects CPUs 31 and 63  new selects 32 33 34 35

+++ src/xenia/base/threading_posix.cc
@@ -1062,7 +1062,10 @@ class PosixCondition<Thread> final : public PosixConditionBase {
     cpu_set_t cpu_set;
     CPU_ZERO(&cpu_set);
     for (auto i = 0u; i < 64; i++) {
-      if (mask & (1 << i)) {
+      // uint64_t(1), not 1: `1 << i` is an int shift, undefined from i == 31
+      // and wrapping to the low 32 bits above it, so every CPU from 32 up was
+      // set from the wrong mask bit.
+      if (mask & (uint64_t(1) << i)) {
         CPU_SET(i, &cpu_set);
       }
     }
